Revert "double, instead of int"
This reverts commit 976c517f38a6fcd785dc1a620d91de4394ee122a.

+++ assemble/compiler.h
@@ -9,7 +9,7 @@ struct cmd {
 	int opcode;
 
 	int types[3];
-	double values[3];
+	int values[3];
 };
 
 struct cmd *assemble_operation(struct operation *op);
